Split xeno-canto api requests into new file.
